+++ src/pages/Gallery/index.tsx
@@ -1,11 +1,12 @@
-import DragCarousel from '@/pages/Gallery/DragCarousel';
+import { CardContent } from '@/components/ui/card';
+import DragCarousel from '@/pages/Gallery/components/DragCarousel';
 
 const Gallery = () => {
   return (
-    <div className="p-0 py-10 text-center">
+    <CardContent className="p-0 py-10 text-center">
       <div className="mb-5 text-xl text-center">갤러리</div>
       <DragCarousel />
-    </div>
+    </CardContent>
   );
 };
 export default Gallery;
