+++ src/pages/HowToCome/HowToCome.tsx
@@ -4,11 +4,13 @@ import Map from '@/pages/HowToCome/components/Map';
 import WayToCome from '@/pages/HowToCome/components/WayToCome';
 
 const HowToCome = () => {
-  return <CardContent className="p-0 py-10 text-center ">
-    <Address />
-    <Map />
-    <WayToCome />
-  </CardContent>;
+  return (
+    <CardContent className="p-0 py-10 text-center ">
+      <Address />
+      <Map />
+      <WayToCome />
+    </CardContent>
+  );
 };
 
 export default HowToCome;
