+++ src/pages/MiddleInfo/components/Greetings/GrettingMessage/index.tsx
