+++ src/pages/MiddleInfo/components/QuoteContent/index.tsx
