+++ src/pages/MiddleInfo/components/Separator/index.tsx
