+++ src/pages/MiddleInfo/index.tsx
