+++ src/pages/TopInfo/components/WeddingTitle.tsx
@@ -1,6 +1,6 @@
 const WeddingTitle = () => {
   return (
-    <h3 className="text-sm uppercase font-wedding serif tracking-[2px] text-[13px] leading-[28px] text-[#898484] ">
+    <h3 className="text-sm uppercase font-wedding serif tracking-[2px] text-[13px] leading-[28px] text-[#898484]">
       THE WEDDING OF
     </h3>
   );
